requirements.txt
supabase
jinja2
python-dotenv